Bug: moved auth inside api/auth
